changes from our Flo experiment

+++ utils/App/StringUtils.h
@@ -12,6 +12,6 @@
 #include <hyendOS/OSVersioning/App/HALVersioning.h>
 #include "connectivity/xprintf/App/xprintf.h"
 
-void getNumberLength(uint16_t*, uint16_t);
+void getNumberLength(uint32_t*, uint32_t);
